backend/pyproject.toml
[tool.poetry]
name = "customyou"
version = "0.1.0"
description = "Record your body health"
authors = ["LazyDragon1123 <lightwalkwithme@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.1.6"
djangorestframework = "^3.14.0"
mysqlclient = "^2.1.1"
django-mysql = "^4.8.0"
django-cors-headers = "^3.13.0"
black = "^23.1.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
autoflake = "^2.0.1"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 119
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
